+++ hw12_13_14_15_calendar/cmd/calendar_scheduler/config.go
@@ -7,18 +7,28 @@ import (
 )
 
 type Config struct {
-	Logger LoggerConf `json:"logger"`
-	AMPQ   AMPQConf   `json:"ampq"`
+	Scheduler SchedulerConfig `json:"scheduler"`
+	DB        DBConf          `json:"db"`
+	Logger    LoggerConf      `json:"logger"`
+	AMPQ      AMPQConf        `json:"ampq"`
 }
 
-type LoggerConf struct {
-	Level string `json:"level"`
-	File  string `json:"file"`
+type SchedulerConfig struct {
+	RecheckDelaySeconds int64 `json:"recheck_delay_seconds"`
+}
+
+type DBConf struct {
+	ConnectionString string `json:"connection_string"`
 }
 
 type AMPQConf struct {
 	URI  string `json:"uri"`
 	Name string `json:"name"`
+}
+
+type LoggerConf struct {
+	Level string `json:"level"`
+	File  string `json:"file"`
 }
 
 func NewConfig() (Config, error) {
@@ -29,8 +39,9 @@ func NewConfig() (Config, error) {
 	}
 
 	return Config{
-			LoggerConf{Level: viper.GetString("logger.level"), File: viper.GetString("logger.file")},
-			AMPQConf{URI: viper.GetString("ampq.uri"), Name: viper.GetString("ampq.name")},
-		},
-		nil
+		SchedulerConfig{RecheckDelaySeconds: viper.GetInt64("scheduler.recheck_delay_seconds")},
+		DBConf{ConnectionString: viper.GetString("db.connection_string")},
+		LoggerConf{Level: viper.GetString("logger.level"), File: viper.GetString("logger.file")},
+		AMPQConf{URI: viper.GetString("ampq.uri"), Name: viper.GetString("ampq.name")},
+	}, nil
 }
