+++ hw12_13_14_15_calendar/cmd/calendar_sender/config.go
@@ -7,28 +7,18 @@ import (
 )
 
 type Config struct {
-	Scheduler SchedulerConfig `json:"scheduler"`
-	DB        DBConf          `json:"db"`
-	Logger    LoggerConf      `json:"logger"`
-	AMPQ      AMPQConf        `json:"ampq"`
-}
-
-type SchedulerConfig struct {
-	RecheckDelaySeconds int64 `json:"recheck_delay_seconds"`
-}
-
-type DBConf struct {
-	ConnectionString string `json:"connection_string"`
-}
-
-type AMPQConf struct {
-	URI  string `json:"uri"`
-	Name string `json:"name"`
+	Logger LoggerConf `json:"logger"`
+	AMPQ   AMPQConf   `json:"ampq"`
 }
 
 type LoggerConf struct {
 	Level string `json:"level"`
 	File  string `json:"file"`
+}
+
+type AMPQConf struct {
+	URI  string `json:"uri"`
+	Name string `json:"name"`
 }
 
 func NewConfig() (Config, error) {
@@ -39,9 +29,8 @@ func NewConfig() (Config, error) {
 	}
 
 	return Config{
-		SchedulerConfig{RecheckDelaySeconds: viper.GetInt64("scheduler.recheck_delay_seconds")},
-		DBConf{ConnectionString: viper.GetString("db.connection_string")},
-		LoggerConf{Level: viper.GetString("logger.level"), File: viper.GetString("logger.file")},
-		AMPQConf{URI: viper.GetString("ampq.uri"), Name: viper.GetString("ampq.name")},
-	}, nil
+			LoggerConf{Level: viper.GetString("logger.level"), File: viper.GetString("logger.file")},
+			AMPQConf{URI: viper.GetString("ampq.uri"), Name: viper.GetString("ampq.name")},
+		},
+		nil
 }
