+++ hw12_13_14_15_calendar/cmd/calendar_sender/main.go
@@ -4,21 +4,19 @@ import (
 	"context"
 	"flag"
 	"fmt"
-	"os"
 	"os/signal"
 	"syscall"
 
+	simpleconsumer "github.com/VladimirButakov/home-work/tree/master/hw12_13_14_15_calendar/internal/amqp/consumer"
 	"github.com/VladimirButakov/home-work/tree/master/hw12_13_14_15_calendar/internal/logger"
-	"github.com/VladimirButakov/home-work/tree/master/hw12_13_14_15_calendar/internal/scheduler"
-	sqlstorage "github.com/VladimirButakov/home-work/tree/master/hw12_13_14_15_calendar/internal/storage/sql"
 	"github.com/VladimirButakov/home-work/tree/master/hw12_13_14_15_calendar/internal/version"
-	_ "github.com/lib/pq"
+	"github.com/streadway/amqp"
 )
 
 var configFile string
 
 func init() {
-	flag.StringVar(&configFile, "config", "/etc/calendar_scheduler/config.json", "Path to configuration file")
+	flag.StringVar(&configFile, "config", "/etc/calendar_sender/config.json", "Path to configuration file")
 }
 
 func main() {
@@ -29,7 +27,7 @@ func main() {
 		return
 	}
 
-	ctx, cancel := context.WithCancel(context.Background())
+	ctx, cancel := signal.NotifyContext(context.Background(), syscall.SIGINT, syscall.SIGTERM)
 	defer cancel()
 
 	config, err := NewConfig()
@@ -39,44 +37,23 @@ func main() {
 
 	logg := logger.New(config.Logger.Level, config.Logger.File)
 
-	storage, err := connectStorage(ctx, config)
+	conn, err := amqp.Dial(config.AMPQ.URI)
 	if err != nil {
-		logg.Error(err.Error())
-
 		panic(err)
 	}
 
-	scheduler := scheduler.NewScheduler(logg, storage, config.Scheduler.RecheckDelaySeconds, config.AMPQ.URI, config.AMPQ.Name)
+	c := simpleconsumer.New(config.AMPQ.Name, conn, logg)
 
-	go func() {
-		signals := make(chan os.Signal, 1)
-		signal.Notify(signals, syscall.SIGINT, syscall.SIGHUP)
-
-		select {
-		case <-ctx.Done():
-			return
-		case <-signals:
-		}
-
-		signal.Stop(signals)
-		cancel()
-	}()
-
-	logg.Info("scheduler is running...")
-
-	scheduler.Start(ctx)
-}
-
-func connectStorage(ctx context.Context, config Config) (*sqlstorage.Storage, error) {
-	storage, err := sqlstorage.New(ctx, config.DB.ConnectionString)
+	msgs, err := c.Consume(ctx, config.AMPQ.Name)
 	if err != nil {
-		return nil, fmt.Errorf("can't create new storage instance, %w", err)
+		logg.Error(fmt.Errorf("cannot consume messages, %w", err).Error())
 	}
 
-	err = storage.Connect(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("can't connect to storage, %w", err)
+	logg.Info("start consuming...")
+
+	for m := range msgs {
+		fmt.Println("receive new message: ", string(m.Data))
 	}
 
-	return storage, nil
+	logg.Info("stopped consuming")
 }
